Transfer uploading functions to 'upload.py' & Add descriptions
See #20
