+++ fanc/upload.py
@@ -1,6 +1,9 @@
+#!/usr/bin/env python3
+
 import numpy as np
 import pandas as pd
 from datetime import datetime
+from caveclient import CAVEclient
 from caveclient.chunkedgraph import root_id_int_list_check
 from cloudvolume.lib import green, red
 from requests.exceptions import HTTPError
@@ -8,15 +11,76 @@ from textwrap import dedent
 from .statebuilder import render_scene
 from .rootID_lookup import segIDs_from_pts_service
 
+CAVE_DATASETS = {'production': 'fanc_production_mar2021',
+                 'sandbox': 'fanc_sandbox'}
+
+class CAVEorganizer(object):
+    """
+    A manager for all the functions that interact with data on CAVE.
+    You need a client to instantiate a organizer:
+
+    client = CAVEclient(datastack_name='my_datastack')
+    organizer = CAVEorganizer(client)
+
+    Then,
+    * organizer.update_soma allows you to upload new manually-detected somas to the soma table
+    """
+    def __init__(
+        self,
+        client,
+        datastack_name=None,
+        server_address=None,
+        auth_token_file=None,
+        auth_token_key=None,
+        auth_token=None,
+        global_only=False,
+        pool_maxsize=None,
+        pool_block=None,
+        desired_resolution=None,
+        info_cache=None
+    ):
+        if client is None:
+            self.client = CAVEclient(datastack_name=datastack_name,
+                                     server_address=server_address,
+                                     auth_token_file=auth_token_file,
+                                     auth_token_key=auth_token_key,
+                                     auth_token=auth_token,
+                                     global_only=global_only,
+                                     pool_maxsize=pool_maxsize,
+                                     pool_block=pool_block,
+                                     desired_resolution=desired_resolution,
+                                     info_cache=info_cache)
+        else:
+            self._client = client
+        self._reset_services()
+
+    def _reset_services(self):
+        self._update_soma = None
+
+    @property
+    def client(self):
+        return self._client
+
+    def get_tables(self, datastack_name=None, version=None):
+        return self._client.materialize.get_tables(datastack_name=datastack_name, 
+                                                   version=version)
+
+    def get_info(self, datastack_name=None):
+        return self._client.info.get_datastack_info(datastack_name=datastack_name)
+
+    @property
+    def update_soma(self):
+        if self._update_soma is None:
+            self._update_soma = SomaTableOrganizer(
+                client=self._client
+            )
+        return self._update_soma
+
+
 def xyz_StringSeries2List(StringSeries: pd.Series):
     pts = StringSeries.str.strip('()').str.split(',',expand=True)
     return pts.astype(int).values.tolist()
 
-def which_label(id, label, dfs):
-    label_list = list(label.values())
-    for i in range(len(label)):
-        if sum(np.isin(dfs[i]["old_nucID"], np.array(id, dtype=np.uint64))) == 1:
-            return label_list[i]
 
 class SomaTableOrganizer(object):
     def __init__(
@@ -28,6 +92,18 @@ class SomaTableOrganizer(object):
     def initialize(self,
                    soma_table_name="",
                    subset_table_name=""):
+        """
+        Set soma table and subset table you want to edit.
+
+        ---Arguments---
+        soma_table_name: string
+            Name of soma table you want to edit. If "" (default), this will
+            be set to the most recent soma table on CAVE: 'soma_mmmYYYY'
+        subset_table_name: string
+            Name of subset table you want to edit. You can choose from either
+            the name of cell types ("neuron", "glia") or the name of subset tables
+            on CAVE: 'celltype_soma_mmmYYYY'
+        """
         if soma_table_name != "":
             self._soma_table_name = soma_table_name
         else:
@@ -126,6 +202,10 @@ class SomaTableOrganizer(object):
             raise UpdateUnsuccessful(red(e.response.text + "Please check after 1 hour. Only works between 10 AM - 12 AM PST."))
 
     def update_tables(self, timestamp=datetime.utcnow()):
+        """
+        Update both soma table and subset table to the most recent version. Will give 
+        an error if newly added points are not ingested yet.
+        """
         self._soma_table = self._check_change(self._soma_table_name, timestamp)
         self._subset_table = self._check_change(self._subset_table_name, timestamp)
 
@@ -144,6 +224,19 @@ class SomaTableOrganizer(object):
             raise UploadUnsuccessful("The information below has root ids that are already registed in soma table. \n {}".format(df.drop(columns=['id'])[overlap].to_string()))
     
     def preview(self, asPoint=True, asSphere=False):
+        """
+        Generate Neuroglancer link to inspect somata on the subset table.
+
+        ---Arguments---
+        asPoint: bool (default True)
+            If True, produce an annotation layer that has all the subset somas as points
+        asSphere: bool (default True)
+            If True, produce an annotation layer that has all the subset somas as sheperes.
+            Their radius are estimated based on their bounding boxes.
+
+        ---Returns---
+        Neuroglancer url (as a string)
+        """
         st = self._soma_table.reindex(columns=['id', 'pt_root_id', 'pt_position', 'bb_start_position', 'bb_end_position'])
 
         annotations = []
@@ -158,6 +251,16 @@ class SomaTableOrganizer(object):
         print(render_scene(annotations=annotations, client=self._client))
 
     def add_dataframe(self, df: pd.DataFrame, bath_size=10000):
+        """
+        Add dataframe to both soma table and subset table.
+
+        ---Arguments---
+        df: pandas.DataFrame
+            A dataframe with coordinates of new somata that you want to upload
+        bath_size: int (default 10000)
+            How many annotations you upload in each batch. The CAVE server does not
+            allow users to upload more than 10000 annotations at a time.
+        """
         print("Checking the format of your table...")
         self.update_tables()
         df_i = df.reset_index(drop=True)
@@ -204,6 +307,7 @@ class SomaTableOrganizer(object):
 class UploadUnsuccessful(Exception):
     def __init__(self, message):      
         super().__init__(red("Uploading failed. ") + message)
+
 
 class UpdateUnsuccessful(Exception):
     def __init__(self, message):
